requirements.txt
requests>=2.28.0
google-cloud-vision
openai
pillow>=9.0.0
matplotlib
plotly
python-dotenv>=0.19.0
streamlit>=1.24.0
opencv-python>=4.5.0
google-generativeai>=0.3.0
groq>=0.3.0
timeout-decorator>=0.5.0
ratelimit>=2.2.1
tenacity>=8.0.0

